requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
Brotli==1.1.0
cffi==1.16.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
nextcord==2.6.0
orjson==3.10.3
pycares==4.4.0
pycparser==2.22
PyNaCl==1.4.0
python-dotenv==1.0.1
six==1.16.0
typing_extensions==4.11.0
yarl==1.9.4
